+++ lib/src/endpoints/me.dart
@@ -10,8 +10,8 @@ class Me extends EndpointPaging {
   Me(SpotifyApiBase api) : super(api);
 
   Future<User> get() async {
-    var jsonString = await _api._get(_path);
-    var map = json.decode(jsonString);
+    final jsonString = await _api._get(_path);
+    final map = json.decode(jsonString);
 
     return User.fromJson(map);
   }
@@ -26,31 +26,79 @@ class Me extends EndpointPaging {
         (json) => Artist.fromJson(json), 'artists', (json) => json);
   }
 
+  /// Check if current user follow the provided artists. The output [bool]
+  /// list is in the same order as the provided artist-id list
+  Future<List<bool>> isFollowing(FollowingType type, List<String> ids) async {
+    assert(ids.isNotEmpty, 'No user/artist id was provided');
+    final jsonString = await _api._get(
+        '$_path/following/contains?type=${type.key}&ids=${ids.join(",")}');
+    final list = List.castFrom<dynamic, bool>(json.decode(jsonString));
+    return list;
+  }
+
+  /// Follow provided users/artists\
+  /// [type] - Type of Follow\
+  /// [ids] - user/artist
+  Future<void> follow(FollowingType type, List<String> ids) async {
+    assert(ids.isNotEmpty, 'No user/artist id was provided');
+    await _api._put("$_path/following?type=${type.key}&ids=${ids.join(",")}");
+  }
+
+  /// Unfollow already following users/artists\
+  /// [type] - Type of Follow\
+  /// [ids] - user/artist
+  Future<void> unfollow(FollowingType type, List<String> ids) async {
+    assert(ids.isNotEmpty, 'No user/artist id was provided');
+    await _api
+        ._delete("$_path/following?type=${type.key}&ids=${ids.join(",")}");
+  }
+
+  /// Get the object currently being played on the user’s Spotify account.
   Future<Player> currentlyPlaying() async {
-    var jsonString = await _api._get('$_path/player/currently-playing');
+    final jsonString = await _api._get('$_path/player/currently-playing');
 
     if (jsonString.isEmpty) {
       return Player();
     }
 
-    var map = json.decode(jsonString);
+    final map = json.decode(jsonString);
     return Player.fromJson(map);
   }
 
-  Future<Iterable<PlayHistory>> recentlyPlayed(
-      {int limit, DateTime after, DateTime before}) async {
+  // Get the currently playing as well as the queued objects.
+  Future<Queue> queue() async {
+    final jsonString = await _api._get('$_path/player/queue');
+
+    if (jsonString.isEmpty) {
+      return Queue();
+    }
+
+    final map = json.decode(jsonString);
+    return Queue.fromJson(map);
+  }
+
+  // Add an object to the queue with a trackId.
+  Future<void> addToQueue(String trackId) async {
+    await _api._post('$_path/player/queue?uri=$trackId');
+  }
+
+  /// Get tracks from the current user’s recently played tracks.
+  /// Note: Currently doesn’t support podcast episodes.
+  CursorPages<PlayHistory> recentlyPlayed(
+      {int? limit, DateTime? after, DateTime? before}) {
     assert(after == null || before == null,
         'Cannot specify both after and before.');
 
-    final jsonString = await _api._get('$_path/player/recently-played?' +
-        _buildQuery({
-          'limit': limit,
-          'after': after?.millisecondsSinceEpoch,
-          'before': before?.millisecondsSinceEpoch
-        }));
-    final map = json.decode(jsonString);
-    return map['items'].map<PlayHistory>((item) => PlayHistory.fromJson(item));
+    return _getCursorPages(
+        '$_path/player/recently-played?' +
+            _buildQuery({
+              'limit': limit,
+              'after': after?.millisecondsSinceEpoch,
+              'before': before?.millisecondsSinceEpoch
+            }),
+        (json) => PlayHistory.fromJson(json));
   }
+
 
   /// Toggle Shuffle For User's Playback.
   ///
@@ -79,30 +127,69 @@ class Me extends EndpointPaging {
     return Player.fromJson(map);
   }
 
+  /// Get the current user's top tracks.
   Future<Iterable<Track>> topTracks() async {
-    var jsonString = await _api._get('$_path/top/tracks');
-    var map = json.decode(jsonString);
+    final jsonString = await _api._get('$_path/top/tracks');
+    final map = json.decode(jsonString);
 
-    var items = map['items'] as Iterable<dynamic>;
+    final items = map['items'] as Iterable<dynamic>;
     return items.map((item) => Track.fromJson(item));
   }
 
+  /// Get the current user's top artists.
   Future<Iterable<Artist>> topArtists() async {
-    var jsonString = await _api._get('$_path/top/artists');
-    var map = json.decode(jsonString);
+    final jsonString = await _api._get('$_path/top/artists');
+    final map = json.decode(jsonString);
 
-    var items = map['items'] as Iterable<dynamic>;
+    final items = map['items'] as Iterable<dynamic>;
     return items.map((item) => Artist.fromJson(item));
   }
 
+  /// Get information about a user’s available devices.
   Future<Iterable<Device>> devices() async {
     return _api._get('$_path/player/devices').then(_parseDeviceJson);
   }
 
-  Iterable<Device> _parseDeviceJson(String jsonString) {
-    var map = json.decode(jsonString);
+  /// Get a list of shows saved in the current Spotify user’s library.
+  Pages<Show> savedShows() {
+    return _getPages('$_path/shows', (json) => Show.fromJson(json['show']));
+  }
 
-    var items = map['devices'] as Iterable<dynamic>;
+  /// gets current user's saved albums in pages
+  Pages<AlbumSimple> savedAlbums() {
+    return _getPages('v1/me/albums', (json) => Album.fromJson(json['album']));
+  }
+
+  /// Save albums for the current-user. It requires the
+  /// `user-library-modify` scope of Spotify WebSDK\
+  /// [ids] - the ids of the albums
+  Future<void> saveAlbums(List<String> ids) async {
+    assert(ids.isNotEmpty, 'No album ids were provided for saving');
+    await _api._put('$_path/albums?ids=${ids.join(",")}');
+  }
+
+  /// Remove albums for the current-user. It requires the
+  /// `user-library-modify` scope of Spotify WebSDK\
+  /// [ids] - the ids of the albums
+  Future<void> removeAlbums(List<String> ids) async {
+    assert(ids.isNotEmpty, 'No album ids were provided for removing');
+    await _api._delete('$_path/albums?ids=${ids.join(",")}');
+  }
+
+  /// Check if passed albums (ids) are saved by current user. The output
+  /// [bool] list is in the same order as the provided album ids list
+  Future<List<bool>> isSavedAlbums(List<String> ids) async {
+    assert(ids.isNotEmpty, 'No album ids were provided for checking');
+    final jsonString =
+        await _api._get('$_path/albums/contains?ids=${ids.join(",")}');
+    final list = List.castFrom<dynamic, bool>(json.decode(jsonString));
+    return list;
+  }
+
+  Iterable<Device> _parseDeviceJson(String jsonString) {
+    final map = json.decode(jsonString);
+
+    final items = map['devices'] as Iterable<dynamic>;
     return items.map((item) => Device.fromJson(item));
   }
 }
@@ -115,4 +202,5 @@ class FollowingType {
   String get key => _key;
 
   static const artist = FollowingType('artist');
+  static const user = FollowingType('user');
 }