+++ lib/src/models/paging.dart
@@ -6,16 +6,14 @@ part of spotify.models;
 typedef ParserFunction<T> = T Function(dynamic object);
 
 Iterable<dynamic> itemsNativeFromJson(List<dynamic> json) => json;
-List<Map> itemsNativeToJson(Iterable<dynamic> items) => List.from(items);
+List<Map> itemsNativeToJson(Iterable<dynamic>? items) =>
+    (items == null) ? [] : List.from(items);
 
-@JsonSerializable(createToJson: false)
-class Paging<T> extends Object {
-  Paging();
-
-  factory Paging.fromJson(Map<String, dynamic> json) => _$PagingFromJson(json);
+class BasePaging<T> extends Object {
+  BasePaging();
 
   /// A link to the Web API endpoint returning the full result of the request.
-  String href;
+  String? href;
 
   /// The requested data
   ///
@@ -23,21 +21,48 @@ class Paging<T> extends Object {
   /// requested data as a deserialized list.
   @JsonKey(
       name: 'items', fromJson: itemsNativeFromJson, toJson: itemsNativeToJson)
-  Iterable<dynamic> itemsNative;
+  Iterable<dynamic>? itemsNative;
 
   /// The maximum number of items in the response (as set in the query or by
   /// default).
-  int limit;
+  int limit = 20;
 
   /// URL to the next page of items. ([null] if none)
-  String next;
+  String? next;
+
+
+}
+
+@JsonSerializable(createToJson: false)
+class Paging<T> extends BasePaging<T> {
+  Paging();
+
+  factory Paging.fromJson(Map<String, dynamic> json) => _$PagingFromJson(json);
 
   /// The offset of the items returned (as set in the query or by default).
-  int offset;
+  int? offset;
 
   /// URL to the previous page of items. (null if none)
-  String previous;
+  String? previous;
 
   /// The total number of items available to return.
-  int total;
+  int total = 0;
+}
+
+@JsonSerializable(createToJson: false)
+class CursorPaging<T> extends BasePaging<T> {
+  CursorPaging();
+
+  factory CursorPaging.fromJson(Map<String, dynamic> json) =>
+      _$CursorPagingFromJson(json);
+
+  Cursor? cursors;
+}
+
+@JsonSerializable(createToJson: false)
+class Cursor extends Object {
+  factory Cursor.fromJson(Map<String, dynamic> json) => _$CursorFromJson(json);
+
+  Cursor();
+  String? after;
 }